ci(workspace): report per-package bundle size in the PR comment [MST-11248]
Extend the per-package coverage PR comment into a combined report
(scripts/coverage-pr-comment.cjs -> pr-package-report.cjs): adds each
package's published size from `npm pack --dry-run` (packed/gzipped + unpacked)
next to coverage. The Build job collects sizes (--collect-sizes) and uploads a
small package-sizes artifact; the comment job downloads it and renders one
unified table. Surfaces oversized packages without a hard gate.

https://uipath.atlassian.net/browse/MST-11248
